fix stock not grabbing name

+++ app/models/value.rb
@@ -4,17 +4,23 @@ class Value < ApplicationRecord
 
   def self.populate_value_table(symbol)
     stock = Stock.find_by_symbol(symbol)
-
-    if Value.find_by_stock_id(stock.id) != nil
-       Value.where(stock_id: stock.id).destroy_all
-    end
-    JSON.parse(Stock::Call.daily(symbol))["Time Series (5min)"].each do |t|
+    JSON.parse(Stock::Call.daily(symbol))["Time Series (15min)"].each do |t|
       if t[0].split[0].to_time > Date.today
         value = Value.new(:price => t[1]["4. close"], :log_time => t[0], :stock_id => stock.id)
         value.save!()
       end
+    end
+  end
 
-
+  def self.populate_value_check(symbol)
+    stock = Stock.find_by_symbol(symbol)
+    if Value.find_by_stock_id(stock.id) != nil
+      if stock.values.first.created_at < Time.now.utc - 500
+        Value.where(stock_id: stock.id).destroy_all
+        Value.populate_value_table(symbol)
+      end
+    else
+      Value.populate_value_table(symbol)
     end
   end
